% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{boot_p_val}
\alias{boot_p_val}
\title{Function for getting p-values out of boot.out object}
\usage{
boot_p_val(boot.out)
}
\arguments{
\item{boot.out}{Bootstrap object returned from the boot function}
}
\value{
Named vector of p-values corresponding to each column of boot.out t-matrix
}
\description{
Function for getting p-values out of boot.out object
}

