% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_se.R
\name{calc_cluster_se}
\alias{calc_cluster_se}
\title{Calculate clustered SEs by sandwich matrix method}
\usage{
calc_cluster_se(X, uhat, clusterby, cluster_ids, G)
}
\arguments{
\item{X}{Vector or matrix of x variables}

\item{uhat}{Vector of residual values from model}

\item{clusterby}{Vector of cluster indices}

\item{cluster_ids}{Vecotr of unique cluster indices}

\item{G}{Integer for the number of groups}
}
\value{
Named vector of clustered standard errors
}
\description{
Calculate clustered SEs by sandwich matrix method
}
\examples{
cluster_se(X = x, uhat = uhat, clusterby = clusterby)
}

