% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_se.R
\name{calc_sandwich_mat}
\alias{calc_sandwich_mat}
\title{Calculate clustered SEs by sandwich matrix method}
\usage{
calc_sandwich_mat(X, bread, uhat, clustervars, cluster_ids, G, comb_n)
}
\arguments{
\item{X}{Matrix of independent variables}

\item{bread}{Bread matrix as X'X}

\item{uhat}{Vector of residual values from model}

\item{cluster_ids}{Vector of unique cluster indices}

\item{G}{Integer for the number of groups}

\item{clusterby}{Vector of cluster indices}
\value{
Named vector of clustered standard errors
}
\description{
Calculate clustered SEs by sandwich matrix method
}

