% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot_tens.R
\name{cluster_meat}
\alias{cluster_meat}
\title{Calculate meat tensor}
\usage{
cluster_meat(clustervars, X, E, k, boot_reps)
}
\arguments{
\item{clustervars}{Vector of cluster indices}

\item{X}{model matrix}

\item{k}{Number of regressors}

\item{boot_reps}{Integer for number of replications}
}
\value{
Meat tensor for all clusters
}
\description{
Calculate meat tensor
}

