% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot_tens.R
\name{cluster_sandwich}
\alias{cluster_sandwich}
\title{Calculate sandiwch tensor}
\usage{
cluster_sandwich(clustervars, X, bread, x_ind, E, k, boot_reps, n, comb_n)
}
\arguments{
\item{clustervars}{Vector of cluster indices}

\item{X}{model matrix}

\item{bread}{Bread matrix}

\item{x_ind}{Integer for index of regressor of interest}

\item{E}{Matrix of residuals}

\item{k}{Number of regressors}

\item{boot_reps}{Integer for number of replications}

\item{n}{Number of observations}

\item{comb_n}{Number of combined cluster dimensions}
}
\value{
Sandwich tensor
}
\description{
Calculate sandiwch tensor
}

