% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_se.R
\name{clustered_se}
\alias{clustered_se}
\title{Calculate clustered standard errors for lm model}
\usage{
clustered_se(data, model, clusterby)
}
\arguments{
\item{data}{Dataframe object containing both model data and clusterby variables}

\item{model}{lm object to calculate clustered standard errors for}

\item{clusterby}{String or fomula indicating the column in `data` to cluster on
Use a string for one-way clustering. a formula to indicate
multiway clustering (e.g. use `~ G + H` to a clustering
using the G and H dimensions}
}
\value{
Named vector of clustered standard errors
}
\description{
This is a flexible function that will calculate cluster-robust
standard errors for an `lm` object. This includes the standard,
one-way clustered standard error (similar to Stata's `cluster` option)
and a "multi-way" clustered standard error, due to Cameron, Gelbach
and Miller (2011)
}
