% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{density_p_val}
\alias{density_p_val}
\title{Calculate density p-value}
\usage{
density_p_val(t, t0, boot_reps)
}
\arguments{
\item{t}{Vector of t-values}

\item{t0}{Original t-value}

\item{boot_reps}{Integer for number of replications}
}
\value{
Bootstrap p-value
}
\description{
Calculate density p-value
}

