% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot_tens.R
\name{eigen_fix}
\alias{eigen_fix}
\title{Spectral decompostion fix}
\usage{
eigen_fix(sandwich)
}
\arguments{
\item{sandwich}{Sandwich matrix}
}
\value{
Corrected matrix
}
\description{
Spectral decompostion fix
}

