% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot_tens.R
\name{gen_boot_weights}
\alias{gen_boot_weights}
\title{Generate bootstrap weight matrix}
\usage{
gen_boot_weights(boot_dist, boot_unique, boot_reps, enum)
}
\arguments{
\item{boot_dist}{Vector of weights for wild bootstrap, string specifying default distribution or a function that only takes the argument "n"}

\item{boot_unique}{Matrix of unique group IDs}

\item{boot_reps}{Number of repititions for resampling data}

\item{enum}{Boolean indicating whether to calculate all possible wild bootstrap combinations, will override boot_reps and report upper and lower bounds}
}
\value{
Matrix of bootstrap weights
}
\description{
Generate bootstrap weight matrix
}

