% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot_tens.R
\name{multiply_cluster_meat}
\alias{multiply_cluster_meat}
\title{Calculate meat tensor for one cluster}
\usage{
multiply_cluster_meat(xe, boot_reps, k)
}
\arguments{
\item{xe}{Combined X and residual matrix}

\item{boot_reps}{Integer for number of replications}

\item{k}{Number of regressors}
}
\value{
Meat tensor for one cluster
}
\description{
Calculate meat tensor for one cluster
}

