% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot_tens.R
\name{p_eval}
\alias{p_eval}
\title{Calculate bootstrap p-value}
\usage{
p_eval(t, t0, boot_reps, bound, absval)
}
\arguments{
\item{t}{Vector of t-values}

\item{t0}{Original t-value}

\item{boot_reps}{Integer for number of replications}

\item{bound}{Boolean indicating whether or not to use bound-MacKinnon correction}

\item{absval}{Boolean indicating whether or not to use absolute valued t-statistics}
}
\value{
Bootstrap p-value
}
\description{
Calculate bootstrap p-value
}

