% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot_tens.R
\name{t_boot_p_val}
\alias{t_boot_p_val}
\title{Calculate bootstrap p-values}
\usage{
t_boot_p_val(se, beta, H0, data, model, clusterby, x_interest, boot_reps,
  absval, bound)
}
\arguments{
\item{se}{Vector of standard errors}

\item{beta}{Vector of coefficients for x of interest}

\item{H0}{Float of integer inticating the null hypothesis, default is 0}

\item{data}{Dataframe of initial data}

\item{model}{lm object of interest}

\item{clusterby}{String or formula indicating variable in data for clustering}

\item{x_interest}{X paramater of interest}

\item{boot_reps}{Integer for number of replications}

\item{absval}{Boolean indicating whether or not to use absolute valued t-statistics}

\item{bound}{Character or character vector indicating which bootstrap p-values to generate}
}
\value{
Bootstrap p-value
}
\description{
Calculate bootstrap p-values
}

