% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot_tens.R
\name{t_wild_cluster_boot}
\alias{t_wild_cluster_boot}
\title{Performs wild clustered bootstrap}
\usage{
t_wild_cluster_boot(data, model, x_interest, clusterby, boot_dist, boot_reps,
  bootby = clusterby, H0 = 0)
}
\arguments{
\item{data}{Dataframe with all data, including group indices}

\item{model}{lm object of interest}

\item{x_interest}{X paramater of interest}

\item{clusterby}{String or formila with name of clusterby variable in data}

\item{boot_dist}{Vector of weights for wild bootstrap or string specifying default distribution}

\item{boot_reps}{Number of repititions for resampling data}

\item{bootby}{String with name of bootby variable in data, default is same as clusterby}

\item{H0}{Float of integer inticating the null hypothesis, default is 0}
}
\value{
p-value corresponding to bootstrap result
}
\description{
Performs wild clustered bootstrap
}

