% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_clust_mle}
\alias{wild_clust_mle}
\title{Generates wild clustered bootstrap list to pass to mle parameter of boot function}
\usage{
wild_clust_mle(model, x_interest, boot_dist, bootby, H0 = 0)
}
\arguments{
\item{model}{lm object of interest}

\item{x_interest}{X paramater of interest}

\item{boot_dist}{Vector of weights for wild bootstrap or string specifying default distribution}

\item{bootby}{String with name of bootby variable in data}

\item{H0}{Integer inticating the alternative hypothesis, default is 0}

\item{clusterby}{Vector of cluster indices corresponding to X variables}
}
\value{
List of specifications to pass to mle option of boot function
}
\description{
Generates wild clustered bootstrap list to pass to mle parameter of boot function
}

