% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_clust_ran}
\alias{wild_clust_ran}
\title{Function for wild clustered bootstrap randomization to pass to ran.gen parameter of boot function}
\usage{
wild_clust_ran(data, mle)
}
\arguments{
\item{data}{Dataframe, which will be the initial data passed to boot data parameter}

\item{mle}{List of wild bootstrap options, generated from gen_wild_mle}
}
\value{
Matrix of y_wild, X variables and clusterby, to be passed to boot statistic function
}
\description{
Function for wild clustered bootstrap randomization to pass to ran.gen parameter of boot function
}

