% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_clust_ran_}
\alias{wild_clust_ran_}
\title{Function to pass to ran.gen parameter of boot function}
\usage{
wild_clust_ran_(data, y_name, uhat, fitted_data, bootby, boot_dist)
}
\arguments{
\item{data}{Dataframe, which will be the initial data passed to boot data parameter}

\item{y_name}{String with name of response variable in data}

\item{uhat}{Vector of residuals from model}

\item{fitted_data}{Vector of fitted data from model}

\item{bootby}{String with name of bootby variable in data}

\item{boot_dist}{Vector of bootstrap distribution}
}
\value{
Randomized data using wild bootstrap proccess
}
\description{
Function to pass to ran.gen parameter of boot function
}

