% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_clust_statistic}
\alias{wild_clust_statistic}
\title{Function for wild clustered bootstrap to pass to statistic parameter of boot function}
\usage{
wild_clust_statistic(data)
}
\arguments{
\item{data}{Dataframe, which will be the data passed from boot ran.gen function at each iteration}
}
\value{
List containing clustered standard error and coefficients for wild bootstrap
}
\description{
Function for wild clustered bootstrap to pass to statistic parameter of boot function
}
\note{
Dataframe must be in order dependent-independent of interest-other independent-cluster
}

