% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_data}
\alias{wild_data}
\title{Return wild bootstrap residuals and fitted values}
\usage{
wild_data(data, model, x_interest, H0)
}
\arguments{
\item{model}{lm object of interest}

\item{x_interest}{X paramater of interest}

\item{H0}{Float of integer inticating the null hypothesis, default is 0}
}
\value{
dataframe with original data and added fitted values and residuals
}
\description{
Return wild bootstrap residuals and fitted values
}

