% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_se}
\alias{wild_se}
\title{Calculate vector of clustered standard errors}
\usage{
wild_se(data_wild, E, X, bread, clusterby, x_ind)
}
\arguments{
\item{data_wild}{Dataframe of original data with wild fitted data and residuals}

\item{E}{Matrix of residuals, which each column representing a bootstrap replication}

\item{X}{Model matrix}

\item{bread}{X*X' matrix}

\item{clusterby}{String or formila with name of clusterby variable in data}

\item{x_ind}{integer indicating the index of x of interest in model matrix}
}
\value{
Vector of standard errors
}
\description{
Calculate vector of clustered standard errors
}

