% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_y}
\alias{wild_y}
\title{Return matrix of wild response variables}
\usage{
wild_y(wild_data, bootby, boot_dist, boot_reps, enum)
}
\arguments{
\item{wild_data}{Dataframe of original data with wild fitted data and residuals}

\item{bootby}{string or formula indicating variables in wild_data to group bootstrap weights by}

\item{boot_dist}{Vector of weights for wild bootstrap, string specifying default distribution or a function that only takes the argument "n"}

\item{boot_reps}{Number of repititions for resampling data}

\item{enum}{Boolean indicating whether to calculate all possible wild bootstrap combinations, will override boot_reps}
}
\value{
Matrix of bootstrap weights
}
\description{
Return matrix of wild response variables
}

