% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_se.R
\name{x_u_cross}
\alias{x_u_cross}
\title{Calculate cross multiplication of cluster X and U matrices}
\usage{
x_u_cross(id, X, uhat, clustervars)
}
\arguments{
\item{uhat}{Vector of residual values from model}

\item{bread}{Bread matrix as X'X}

\item{clusterby}{Vector of cluster indices}

\item{cluster_ids}{Vector of unique cluster indices}

\item{G}{Integer for the number of groups}
}
\value{
Named vector of clustered standard errors
}
\description{
Calculate cross multiplication of cluster X and U matrices
}

