% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_se.R
\name{clustered_se}
\alias{clustered_se}
\title{Calculate clustered standard errors for lm model}
\usage{
clustered_se(data, model, clusterby, G = NULL, cluster_ids = NULL)
}
\arguments{
\item{data}{Dataframe containing both model data and clusterby variables}

\item{model}{lm object for which clustered errors sought}

\item{clusterby}{Vector of cluster indices corresponding to X variables}

\item{G}{Total number of groups, should be specified for performance}

\item{cluster_ids}{Vector of unique cluster_ids, should be specified for performance}
}
\value{
Named vector of clustered standard errors
}
\description{
Calculate clustered standard errors for lm model
}

