% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{curry_wild_clust_ran}
\alias{curry_wild_clust_ran}
\title{Function to pass to ran.gen parameter of boot function}
\usage{
curry_wild_clust_ran(data, y_name, uhat, fitted_data, bootby, boot_ids, bG,
  boot_dist)
}
\arguments{
\item{data}{Dataframe, which will be the initial data passed to boot data parameter}

\item{y_name}{String with name of response variable}

\item{uhat}{Vector of residuals from model}

\item{fitted_data}{Vector of fitted data from model}

\item{bootby}{Vector of group ids for use in wild bootstrap}

\item{boot_ids}{Vector containing just the unique group ids}

\item{bG}{Integer indicating the total number of unique groups}
}
\value{
Matrix of y_wild, X variables and clusterby, to be passed to boot statistic function
}
\description{
Function to pass to ran.gen parameter of boot function
}

