% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{curry_wild_clust_statistic}
\alias{curry_wild_clust_statistic}
\title{Function for wild clustered bootstrap to pass to statistic parameter of boot function}
\usage{
curry_wild_clust_statistic(data, form, x_interest, clusterby)
}
\arguments{
\item{data}{Dataframe, which will be the data passed from boot ran.gen function at each iteration}

\item{x_interest}{String with name of the independent variable of interest}

\item{clusterby}{Vector of group ids for clustering}

\item{formula}{Formula object to be used on data}
}
\value{
List containing clustered standard error and coefficients for wild bootstrap
}
\description{
Function for wild clustered bootstrap to pass to statistic parameter of boot function
}

