% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_clust_mle}
\alias{wild_clust_mle}
\title{Generates wild clustered bootstrap list to pass to mle parameter of boot function}
\usage{
wild_clust_mle(model, param, clusterby, boot_dist, bootby = clusterby,
  H0 = 0)
}
\arguments{
\item{model}{lm object of interest}

\item{param}{X paramater of interest}

\item{clusterby}{Vector of cluster indices corresponding to X variables}

\item{boot_dist}{Vector of weights for wild bootstrap}

\item{bootby}{Vector of bootstrap group indices corresponding to X variables, default is same as clusterby}

\item{H0}{Integer inticating the alternative hypothesis, default is 0}
}
\value{
List of specifications to pass to mle option of boot function
}
\description{
Generates wild clustered bootstrap list to pass to mle parameter of boot function
}

