% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_clust_ran}
\alias{wild_clust_ran}
\title{Function to curry function for ran.gen parameter of boot function}
\usage{
wild_clust_ran(data, mle)
}
\arguments{
\item{data}{Data of interest, passed by boot function}

\item{mle}{List of optional parameters, passed by boot function}
}
\value{
Randomized data using wild bootstrap proccess
}
\description{
Function to curry function for ran.gen parameter of boot function
}

