% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_clust_ran}
\alias{wild_clust_ran}
\title{Function to curry function for ran.gen parameter of boot function}
\usage{
wild_clust_ran(model, x_interest, bootby, boot_dist, H0 = 0)
}
\arguments{
\item{model}{lm model of interest}

\item{x_interest}{String with name of independent variable of interest}

\item{bootby}{String with name of bootby variable in data}

\item{boot_dist}{Vector of bootstrap distribution}

\item{H0}{Number indicating the null hypothesis, default is 0}
}
\value{
Function to pass to ran.gen parameter of boot function
}
\description{
Function to curry function for ran.gen parameter of boot function
}

