% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_clust_statistic}
\alias{wild_clust_statistic}
\title{Function to curry function to pass to statistic parameter of boot function}
\usage{
wild_clust_statistic(model, x_interest, clusterby)
}
\arguments{
\item{model}{lm model of interest}

\item{x_interest}{String with name of the independent variable of interest}

\item{clusterby}{String or list of strings with clusterby variable name or names}
}
\value{
List containing clustered standard error and coefficients for wild bootstrap
}
\description{
Function to curry function to pass to statistic parameter of boot function
}

