% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_cluster_boot.R
\name{wild_clust_statistic}
\alias{wild_clust_statistic}
\title{Function for wild clustered bootstrap to pass to statistic parameter of boot function}
\usage{
wild_clust_statistic(data, model, x_interest, clusterby, H0)
}
\arguments{
\item{data}{Dataframe, which will be the data passed from boot ran.gen function at each iteration}

\item{model}{lm model of interest}

\item{x_interest}{String with name of the independent variable of interest}

\item{clusterby}{String with name of clusterby variable in data}

\item{H0}{Number indicating the null hypothesis, default is 0}
}
\value{
List containing clustered standard error and coefficients for wild bootstrap
}
\description{
Function for wild clustered bootstrap to pass to statistic parameter of boot function
}

